third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/shadowroot-type-expected.txt

CONSOLE ERROR: line 15: Uncaught TypeError: host.createShadowRoot is not a function
This test ensures ShadowRootType is correctly reflected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
